Update dependency scope for DB backends
When upgrading a project from Quartz 2.3.2 to 2.5.0 we ran into the following error:

```
Exception in thread "main" java.lang.NoClassDefFoundError: com/mchange/v2/c3p0/ComboPooledDataSource
	at org.quartz.utils.C3p0PoolingConnectionProvider.initialize(C3p0PoolingConnectionProvider.java:187)
	at org.quartz.utils.C3p0PoolingConnectionProvider.<init>(C3p0PoolingConnectionProvider.java:106)
	at com.rms.mapping.system.scheduler.QuartzSchedulerManager$.init(QuartzSchedulerManager.scala:225)
	at com.rms.mapping.system.scheduler.QuartzSchedulerManager$.apply(QuartzSchedulerManager.scala:201)
```

Looking at the Gradle build, it look like the `c3p0` dependency is declared as `implementation` which means it's not exposed on the runtime class path. While the workaround from #1310 is sufficient if these types are exposed in the library API then it seems to make sense to change their declaration scope.

 This PR updates the dependency declaration to restore the behavior of the 2.3.x library.

If these dependencies were intentionally hidden then perhaps a comment in the release notes or the README would be helpful?

Signed-off-by: Nick Robison <nickrobison@users.noreply.github.com>